Spread the Truth!
Nearly 150 million Russians do not know the truth about the causes or course of the war in Ukraine. It is fed with the lies of the Kremlin propaganda. There is no free media in Russia and the internet is censored. But each of us has the ability to send a direct message to the citizens of this enslaved country.